requirements.txt
# Poetry to Image - Dependencies
# Python 3.11+

# GUI Framework
PySide6>=6.6.0

# Data Validation
pydantic>=2.0.0

# HTTP Client
requests>=2.31.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dateutil>=2.8.0
